ai_service/backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
python-dotenv==1.2.2
cloudinary==1.40.0
google-genai==1.66.0
google-generativeai==0.8.6
TTS==0.22.0
torch==2.1.2
torchaudio==2.1.2
transformers==4.37.2
numpy
scipy
librosa
soundfile
tqdm